Moved `mod deref` and `fn rv_place` to `mod mir_utils` (`fn rv_place` is also made `pub`).
