+++ dynamic_instrumentation/src/mir_utils/deref.rs
