+++ dynamic_instrumentation/src/mir_utils/mod.rs
@@ -1,0 +1,25 @@
+use rustc_middle::mir::{Place, Rvalue};
+
+pub mod deref;
+
+/// Get the one and only input [`Place`], if applicable.
+pub fn rv_place<'tcx>(rv: &'tcx Rvalue) -> Option<Place<'tcx>> {
+    use Rvalue::*;
+    match rv {
+        Use(op) => op.place(),
+        Repeat(op, _) => op.place(),
+        Ref(_, _, p) => Some(*p),
+        // ThreadLocalRef
+        AddressOf(_, p) => Some(*p),
+        Len(p) => Some(*p),
+        Cast(_, op, _) => op.place(),
+        // BinaryOp
+        // CheckedBinaryOp
+        // NullaryOp
+        UnaryOp(_, op) => op.place(),
+        Discriminant(p) => Some(*p),
+        // Aggregate
+        ShallowInitBox(op, _) => op.place(),
+        _ => None,
+    }
+}
