+++ dynamic_instrumentation/src/point/cast.rs
@@ -10,7 +10,7 @@ use rustc_span::DUMMY_SP;
 
 use crate::{
     arg::{ArgKind, InstrumentationArg},
-    deref::remove_outer_deref,
+    mir_utils::deref::remove_outer_deref,
 };
 
 /// Cast an argument from pointer to `usize`, if needed.
