+++ dynamic_instrumentation/src/point/source.rs
@@ -1,7 +1,7 @@
 use c2rust_analysis_rt::mir_loc::MirPlace;
 use rustc_middle::mir::{Local, Operand, Place, Rvalue};
 
-use crate::util::Convert;
+use crate::{mir_utils::rv_place, util::Convert};
 
 pub trait Source {
     fn source(&self) -> Option<MirPlace>;
@@ -41,27 +41,5 @@ impl Source for Local {
 impl Source for u32 {
     fn source(&self) -> Option<MirPlace> {
         Local::from_u32(*self).source()
-    }
-}
-
-/// Get the one and only input [`Place`], if applicable.
-fn rv_place<'tcx>(rv: &'tcx Rvalue) -> Option<Place<'tcx>> {
-    use Rvalue::*;
-    match rv {
-        Use(op) => op.place(),
-        Repeat(op, _) => op.place(),
-        Ref(_, _, p) => Some(*p),
-        // ThreadLocalRef
-        AddressOf(_, p) => Some(*p),
-        Len(p) => Some(*p),
-        Cast(_, op, _) => op.place(),
-        // BinaryOp
-        // CheckedBinaryOp
-        // NullaryOp
-        UnaryOp(_, op) => op.place(),
-        Discriminant(p) => Some(*p),
-        // Aggregate
-        ShallowInitBox(op, _) => op.place(),
-        _ => None,
     }
 }
